fix: address review findings for logs route and chat pages
- Revert innerJoin back to leftJoin for agentComposeVersions/agentComposes
  in logs route to prevent silently dropping runs with null compose refs
- Use agentRuns.orgId instead of agentComposes.orgId for org-scoping
  (safer when compose version is deleted via onDelete: set null)
- Extract duplicated model preference logic into shared
  zero-model-preference.ts (useModelSelection hook)
- Remove copy-pasted code from zero-chat-page.tsx and
  zero-session-chat-page.tsx
